feat(benchmark): add io saver
to pro mq is worked
