+++ util/file_impl.cc
@@ -49,4 +49,108 @@ FileImpl::~FileImpl() {
   delete sink_;
   delete source_;
 }
+
+Status SpecialEnv::NewWritableFile(const std::string& f, WritableFile** r) {
+  class DataFile : public WritableFile {
+   private:
+    SpecialEnv* const env_;
+    WritableFile* const base_;
+    const std::string fname_;
+
+   public:
+    DataFile(SpecialEnv* env, WritableFile* base, const std::string& fname)
+        : env_(env), base_(base), fname_(fname) {}
+
+    ~DataFile() { delete base_; }
+    Status Append(const Slice& data) {
+      if (env_->no_space_.load(std::memory_order_acquire)) {
+        // Drop writes on the floor
+        return Status::OK();
+      } else {
+        return base_->Append(data);
+      }
+    }
+    Status Close() {
+      Status s = base_->Close();
+      if (s.ok() && IsLogFile(fname_) &&
+          env_->log_file_close_.load(std::memory_order_acquire)) {
+        s = Status::IOError("simulated log file Close error");
+      }
+      return s;
+    }
+    Status Flush() { return base_->Flush(); }
+    Status Sync() {
+      if (env_->data_sync_error_.load(std::memory_order_acquire)) {
+        return Status::IOError("simulated data sync error");
+      }
+      while (env_->delay_data_sync_.load(std::memory_order_acquire)) {
+        DelayMilliseconds(100);
+      }
+      return base_->Sync();
+    }
+  };
+  class ManifestFile : public WritableFile {
+   private:
+    SpecialEnv* env_;
+    WritableFile* base_;
+
+   public:
+    ManifestFile(SpecialEnv* env, WritableFile* b) : env_(env), base_(b) {}
+    ~ManifestFile() { delete base_; }
+    Status Append(const Slice& data) {
+      if (env_->manifest_write_error_.load(std::memory_order_acquire)) {
+        return Status::IOError("simulated writer error");
+      } else {
+        return base_->Append(data);
+      }
+    }
+    Status Close() { return base_->Close(); }
+    Status Flush() { return base_->Flush(); }
+    Status Sync() {
+      if (env_->manifest_sync_error_.load(std::memory_order_acquire)) {
+        return Status::IOError("simulated sync error");
+      } else {
+        return base_->Sync();
+      }
+    }
+  };
+
+  if (non_writable_.load(std::memory_order_acquire)) {
+    return Status::IOError("simulated write error");
+  }
+
+  Status s = target()->NewWritableFile(f, r);
+  if (s.ok()) {
+    if (IsLdbFile(f) || IsLogFile(f)) {
+      *r = new DataFile(this, *r, f);
+    } else if (IsManifestFile(f)) {
+      *r = new ManifestFile(this, *r);
+    }
+  }
+  return s;
+}
+
+Status SpecialEnv::NewRandomAccessFile(const std::string& f, RandomAccessFile** r) {
+  class CountingFile : public RandomAccessFile {
+   private:
+    RandomAccessFile* target_;
+    AtomicCounter* counter_;
+
+   public:
+    CountingFile(RandomAccessFile* target, AtomicCounter* counter)
+        : target_(target), counter_(counter) {}
+    ~CountingFile() override { delete target_; }
+    Status Read(uint64_t offset, size_t n, Slice* result,
+                char* scratch) const override {
+      counter_->Increment();
+      return target_->Read(offset, n, result, scratch);
+    }
+  };
+
+  Status s = target()->NewRandomAccessFile(f, r);
+  if (s.ok() && count_random_reads_) {
+    *r = new CountingFile(*r, &random_read_counter_);
+  }
+  return s;
+}
 };  // namespace leveldb
