+++ util/file_impl.h
@@ -7,11 +7,16 @@
 #ifndef LEVELDB_FILE_IMPL_H
 #define LEVELDB_FILE_IMPL_H
 
+#include <atomic>
+
 #include "leveldb/env.h"
 #include "leveldb/status.h"
 
+#include "port/thread_annotations.h"
 #include "table/format.h"
 #include "util/crc32c.h"
+
+#include "mutexlock.h"
 
 namespace leveldb {
 
@@ -74,6 +79,92 @@ class FileImpl {
   StringSink* sink_;
   StringSource* source_;
   uint64_t write_offset_;
+};
+
+class AtomicCounter {
+ public:
+  AtomicCounter() : count_(0) {}
+  void Increment() { IncrementBy(1); }
+  void IncrementBy(int count) LOCKS_EXCLUDED(mu_) {
+    MutexLock l(&mu_);
+    count_ += count;
+  }
+  int Read() LOCKS_EXCLUDED(mu_) {
+    MutexLock l(&mu_);
+    return count_;
+  }
+  void Reset() LOCKS_EXCLUDED(mu_) {
+    MutexLock l(&mu_);
+    count_ = 0;
+  }
+
+ private:
+  port::Mutex mu_;
+  int count_ GUARDED_BY(mu_);
+};
+
+namespace {
+void DelayMilliseconds(int millis) {
+  Env::Default()->SleepForMicroseconds(millis * 1000);
+}
+
+bool IsLdbFile(const std::string& f) {
+  return strstr(f.c_str(), ".ldb") != nullptr;
+}
+
+bool IsLogFile(const std::string& f) {
+  return strstr(f.c_str(), ".log") != nullptr;
+}
+
+bool IsManifestFile(const std::string& f) {
+  return strstr(f.c_str(), "MANIFEST") != nullptr;
+}
+}  // namespace
+
+// Special Env used to delay background operations.
+class SpecialEnv : public EnvWrapper {
+ public:
+  // For historical reasons, the std::atomic<> fields below are currently
+  // accessed via acquired loads and release stores. We should switch
+  // to plain load(), store() calls that provide sequential consistency.
+
+  // sstable/log Sync() calls are blocked while this pointer is non-null.
+  std::atomic<bool> delay_data_sync_;
+
+  // sstable/log Sync() calls return an error.
+  std::atomic<bool> data_sync_error_;
+
+  // Simulate no-space errors while this pointer is non-null.
+  std::atomic<bool> no_space_;
+
+  // Simulate non-writable file system while this pointer is non-null.
+  std::atomic<bool> non_writable_;
+
+  // Force sync of manifest files to fail while this pointer is non-null.
+  std::atomic<bool> manifest_sync_error_;
+
+  // Force write to manifest files to fail while this pointer is non-null.
+  std::atomic<bool> manifest_write_error_;
+
+  // Force log file close to fail while this bool is true.
+  std::atomic<bool> log_file_close_;
+
+  bool count_random_reads_;
+  AtomicCounter random_read_counter_;
+
+  explicit SpecialEnv(Env* base)
+      : EnvWrapper(base),
+        delay_data_sync_(false),
+        data_sync_error_(false),
+        no_space_(false),
+        non_writable_(false),
+        manifest_sync_error_(false),
+        manifest_write_error_(false),
+        log_file_close_(false),
+        count_random_reads_(false) {}
+  Status NewWritableFile(const std::string& f, WritableFile** r);
+
+  Status NewRandomAccessFile(const std::string& f, RandomAccessFile** r);
 };
 
 }  // namespace leveldb
